Have content hash as long
Allows clients of the library to implement functions with less collision
chances.

+++ Animations/src/main/java/io/doist/recyclerviewext/animations/AnimatedAdapter.java
@@ -21,7 +21,7 @@ public abstract class AnimatedAdapter<VH extends RecyclerView.ViewHolder>
     /**
      * @see DataSetDiffer.Callback#getItemContentHash(int)
      */
-    public abstract int getItemContentHash(int position);
+    public abstract long getItemContentHash(int position);
 
     /**
      * Returns whether animations are enabled or not.
